Update README.md
I completed this coin flip experiment a little faster than I expected, and on Super Bowl Sunday no less. I remembered a lesson from one of my statistics classes that a coin flip is never exactly 50/50, so I decided to go down the internet rabbit hole to re-work the code to be a little closer to a real-life coin flip. 

This being Super Bowl Sunday, there were pages and pages of articles analyzing the different variables of Super Bowl coin flips. Analysts provided historical data, weights and diameters, and coin flip win/loss stats for the two super bowl teams to help bettors pick a bet for tonight’s coin toss. All of this seemed irrelevant to me. Additionally, Super Bowl coin flips are an especially small sample size to draw conclusions from, and the coin wasn’t even made by the same mint until 1994. If you want to place a bet on the coin toss this year, don’t read any of these articles. None of them will provide you with the right variables to make a bet in your favor. For my purposes, none of these articles could provide me with what I needed to make my experiment closer to a real-life coin flip.

One variable that has been shown to be relevant is which side is facing up whenever a coin is tossed. According to Persi Diaconis of Stanford, whose study was quoted in several articles including this one, the odds of a coin flip are 51-49 percent, favoring the side that started facing up. This is a variable I can account for. I re-named my variables to make it clear that “heads” or “tails” isn’t what matters, but whichever side is “face up” or “face down” when it starts.  I changed my code to favor one variable over another by having the program choose a random number between 1-100, instead of giving it a binary choice. If the selected number was less than 52, then the hypothetical coin landed with the same side facing up that it started with. If the number was more than that, that the face-up side landed facing down.  

Predictably, as I observed from running my code a few times, the variable that is favored, “wins” the most. I suppose if I assume one variable to always be face up at the start, (For example “heads”), then a streak of 6 of either variable is more likely to occur in real-life than it is in the original experiment. I’m not entirely confident in my conclusion, so I will revisit this once I revisit my stats class notes. 

Key take aways from the experiments: If you have to place your bet before the coin is placed to be flipped, your odds are still 50/50 of getting it right. If you can see which side of the coin is face up before it is flipped, your odds are 51/49 in favor of the face up side.